feat: tratamento de database error

+++ src/main/java/com/projdn/projdn/services/UserService.java
@@ -4,10 +4,12 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.projdn.projdn.entities.User;
 import com.projdn.projdn.repositories.UserRepository;
+import com.projdn.projdn.resources.exceptions.DatabaseException;
 import com.projdn.projdn.services.exceptions.ResourceNotFoundException;
 
 @Service
@@ -23,17 +25,27 @@ public class UserService {
 	public User findById(Long id) {
 
 		Optional<User> obj = userRepository.findById(id);
-		return obj.orElseThrow(()-> new ResourceNotFoundException(id));
+		return obj.orElseThrow(() -> new ResourceNotFoundException(id));
 	}
 
 	public User insert(User obj) {
 		return userRepository.save(obj);
 	}
-	
+
 	public void delete(Long id) {
-		 userRepository.deleteById(id);
+		try {
+			
+		if (userRepository.existsById(id)) {
+			userRepository.deleteById(id);
+		} else {
+			throw new ResourceNotFoundException(id);
+		}
+		}catch(DataIntegrityViolationException e) {
+			throw new DatabaseException(e.getMessage());
+		}
+
 	}
-	
+
 	public User update(Long id, User obj) {
 		User entity = userRepository.getReferenceById(id);
 		updateData(entity, obj);
@@ -45,5 +57,5 @@ public class UserService {
 		entity.setEmail(obj.getEmail());
 		entity.setPhone(obj.getPhone());
 	}
-	
+
 }
